int debug

+++ simple_gram_desktop/constructor_app/widgets/login_widget.py
@@ -154,8 +154,8 @@ class LoginWidget(QWidget):
     def resizeEvent(self, event: PySide6.QtGui.QResizeEvent) -> None:
         # toDo: If this will be used in the future, then put the colors in the parameters
         self._background = QLinearGradient()
-        self._background.setStart(QPoint(self.width()/2, 0))
-        self._background.setFinalStop(QPoint(self.width()/2, self.height()))
+        self._background.setStart(QPoint(int(self.width()/2), 0))
+        self._background.setFinalStop(QPoint(int(self.width()/2), self.height()))
         self._background.setColorAt(0, QColor(57, 178, 146))
         self._background.setColorAt(0.5, QColor(68, 159, 167))
         self._background.setColorAt(1, QColor(82, 136, 193))
